deps: update V8 to 8.7.220

+++ deps/v8/src/objects/internal-index.h
@@ -40,7 +40,8 @@ class InternalIndex {
     return static_cast<uint32_t>(entry_);
   }
   constexpr int as_int() const {
-    CONSTEXPR_DCHECK(entry_ <= std::numeric_limits<int>::max());
+    CONSTEXPR_DCHECK(entry_ <=
+                     static_cast<size_t>(std::numeric_limits<int>::max()));
     return static_cast<int>(entry_);
   }
